refactor: member 패키지 생성
