+++ src/main/java/project/myblog/auth/dto/LoginMember.java
@@ -1,6 +1,6 @@
 package project.myblog.auth.dto;
 
-import project.myblog.domain.Member;
+import project.myblog.domain.member.Member;
 
 import java.util.Objects;
