+++ src/main/java/project/myblog/repository/MemberRepository.java
@@ -1,7 +1,7 @@
 package project.myblog.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import project.myblog.domain.Member;
+import project.myblog.domain.member.Member;
 
 import java.util.Optional;
