+++ src/test/java/project/myblog/unit/member/MemberTest.java
@@ -1,7 +1,7 @@
 package project.myblog.unit.member;
 
 import org.junit.jupiter.api.Test;
-import project.myblog.domain.Member;
+import project.myblog.domain.member.Member;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static project.myblog.acceptance.member.MemberStepsRequest.EMAIL;
